requirements.txt
# Core dependencies
chonkie>=0.3.0
openai>=1.0.0
qdrant-client>=1.7.0
streamlit>=1.40.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
pypdf>=5.0.0
python-dotenv>=1.0.0

# Additional utilities
reportlab>=4.2.0  # For generating mock PDFs
PyPDF2>=3.0.0
